% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betafunctions.R
\name{Beta.gfx.poly.pdf}
\alias{Beta.gfx.poly.pdf}
\title{Coordinate Generation for Marking an Area Under the Curve for the Beta Probability Density Distribution.}
\usage{
Beta.gfx.poly.pdf(from, to, by, alpha, beta, l = 0, u = 1)
}
\arguments{
\item{from}{The point of the \code{x}-axis from where to start producing \code{y}-density values.}

\item{to}{The point of the x-axis to where y-density values are to be produced.}

\item{by}{The resolution (or spacing) at which to produce y-density values.}

\item{alpha}{The alpha (first) shape-parameter value for the Standard Beta probability density distribution.}

\item{beta}{The beta (second) shape-parameter for the Standard Beta probability density distribution.}

\item{l}{The lower-bound location parameter of the Beta distribution.}

\item{u}{The upper-bound location parameter of the Beta distribution.}
}
\value{
A two-column matrix with density-values of y to plot against corresponding location values of x.
}
\description{
Plotting tool, producing a two-column matrix with values of \code{y} corresponding to locations on \code{x}. Useful for shading areas under the curve when tracing the line for the Beta probability density functions.
}
\examples{
# To box in an area under a four-parameter Beta distribution with location
# parameters l = .25 and u = .75, and shape parameters alpha = 5 and
# rbeta = 3, from 0.4 to 0.6:
plot(NULL, xlim = c(0, 1), ylim = c(0, 7))
coords <- Beta.gfx.poly.pdf(from = 0.4, to = 0.6, by = 0.001, alpha = 5,
beta = 3, l = 0.25, u = 0.75)
polygon(coords)
}
